api/pyproject.toml
[tool.poetry]
name = "api"
version = "0.1.0"
description = ""
authors = ["danielAdama <adamadaniel321@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
flask = "^3.1.0"
groq = "^0.12.0"
langchain-groq = "^0.2.1"
langchain-core = "^0.3.18"
python-dotenv = "^1.0.1"
flask-cors = "^5.0.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
